// packages/plugins/blog/src/models/types.ts
export interface PostCreateData {
  name: string;
  description?: string;
  content?: string;
  status?: string;
  authorId?: number;
  isFeatured?: boolean;
  image?: string;
  formatType?: string;
  translationGroupId?: string;
  categoryIds?: number[];
  tagIds?: number[];
  meta?: Record<string, any>;
}

export interface PostUpdateData {
  name?: string;
  description?: string;
  content?: string;
  status?: string;
  isFeatured?: boolean;
  image?: string;
  formatType?: string;
  categoryIds?: number[];
  tagIds?: number[];
  meta?: Record<string, any>;
}

export interface CategoryCreateData {
  name: string;
  parentId?: number;
  description?: string;
  status?: string;
  authorId?: number;
  icon?: string;
  order?: number;
  isFeatured?: boolean;
  isDefault?: boolean;
}

export interface CategoryUpdateData {
  name?: string;
  parentId?: number;
  description?: string;
  status?: string;
  icon?: string;
  order?: number;
  isFeatured?: boolean;
  isDefault?: boolean;
}

export interface TagCreateData {
  name: string;
  authorId?: number;
  description?: string;
  status?: string;
}